Fix coercion of large integer-y numbers to string for xls; fixes #346 (#405)
* Add failing test re: coercing large integer-y numbers to string in xls

* Cast to a longer integer during string coercion; fixes #346

* Trying `long long`

+++ src/XlsCell.h
@@ -245,7 +245,7 @@ public:
       // examples: social security or student number
       double intpart;
       if (std::modf(cell_->d, &intpart) == 0.0) {
-        strs << std::fixed << (int)cell_->d;
+        strs << std::fixed << (long long)cell_->d;
       } else {
         strs << cell_->d;
       }
